+++ pyinaturalist/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = 'Nicolas Noé'
 __email__ = 'nicolas@niconoe.eu'
-__version__ = '0.7.0'
+__version__ = '0.8.0'
 
 DEFAULT_USER_AGENT = 'Pyinaturalist/{version}'.format(version=__version__)